Worked on resource importing

+++ src/modloader/classes/components/Armor.java
@@ -1,5 +1,5 @@
 package modloader.classes.components;
 
 public class Armor {
-    public float resistance;
+    public double resistance;
 }
