+++ src/modloader/classes/components/Axe.java
@@ -1,5 +1,5 @@
 package modloader.classes.components;
 
 public class Axe {
-    public float efficiency;
+    public double efficiency;
 }
