+++ src/modloader/classes/components/Chest.java
@@ -1,5 +1,5 @@
 package modloader.classes.components;
 
 public class Chest {
-    public float watertightness;
+    public double watertightness;
 }
