+++ src/modloader/classes/components/Dapperness.java
@@ -1,5 +1,5 @@
 package modloader.classes.components;
 
 public class Dapperness {
-    public float rate;
+    public double rate;
 }
