+++ src/modloader/classes/components/Edible.java
@@ -3,7 +3,7 @@ package modloader.classes.components;
 public class Edible {
     public boolean isEnabled;
 
-    public float health;
-    public float sanity;
-    public float hunger;
+    public double health;
+    public double sanity;
+    public double hunger;
 }
