+++ src/modloader/classes/components/Hand.java
@@ -1,5 +1,5 @@
 package modloader.classes.components;
 
 public class Hand {
-    public float watertightness;
+    public double watertightness;
 }
