+++ src/modloader/classes/components/Hat.java
@@ -1,5 +1,5 @@
 package modloader.classes.components;
 
 public class Hat {
-    public float watertightness;
+    public double watertightness;
 }
