+++ src/modloader/classes/components/Pickaxe.java
@@ -1,5 +1,5 @@
 package modloader.classes.components;
 
 public class Pickaxe {
-    public float efficiency;
+    public double efficiency;
 }
